refactor: TeamRankResponse 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/TeamRankResponse.kt
@@ -6,6 +6,6 @@ data class TeamRankResponse(
     val id: UUID,
     val teamName: String?,
     val schoolName: String,
-    val totalMoney: Int,
+    val totalMoney: Long,
     val isMyTeam: Boolean
 )
